refactor: restore and use document dropdown reindex
